refactor(compiler): split lower into TIR-mutate + planner + TIR→NIR translator (Phase 1: value_copy)
Restructure the lower phase exit so it reads as a translation function
instead of a TIR mutate + tail copy:

  in-place TIR mutation → lower::plan::plan(&mut flat) → lower::translate::translate(flat, plan)

Phase 1 migrates `value_copy` to this shape:

- `lower::plan::value_copy::plan` runs `insert::insert_value_copy_calls`
  (TIR-mutating) and `synthesize::synthesize_helpers` (which pushes the
  `$value_copy$T<id>` helpers into `flat.functions` and returns a
  `TypeId → (ModuleSource, name)` map).
- The translator's `Call` arm rewrites `builtin::copy_value::<T>(x)`
  markers — placed by the planner — into direct calls to the helpers
  using `plan.value_copy.name_for_type`. Both user bodies and the
  synthesized helper bodies (which themselves contain
  `copy_value::<NestedT>(...)` for nested fields) pick up the rewrite
  uniformly through the TIR → NIR fold.

`nir_convert::flat_to_nir` is moved to `lower::translate::translate` and
gains a `plan: LowerPlan` parameter. All convert_* recursive functions
that touch expressions become methods on a `Translator` struct holding
`&LowerPlan`; pure data converters (binary ops, type params, etc.) stay
as free functions.

Subsequent phases will migrate string / closure / boxing / globals /
pattern / wide_int into the same planner + translator split, after
which the in-place chain disappears entirely and TIR-side-only
variants (e.g. `ClosureToCanonical`) can be dropped.

See `docs/wep-2026-05-11-nir.md`.
